requirements.txt
flask==3.1.1
yt-dlp==2025.3.21
python-json-logger==3.3.0
qrcode==8.2
pillow==12.1.1
